Fix the no longer tested Py_LIMITED_API
The limited API has always been tested in COIN.

Not so last time when I checked this.
I had a check-in which defined PyImport_GetModule for Python
versions < 3.7, but did not see that this function does not
exist when the Py_LIMITED_API is defined.

The check-in should therefore have been broken, but it went
through. That means that the Py_LIMITED_API is not tested,
at least since commit c82ec2bc on 2020-05-18.

Should we change the default, and how?

Change-Id: I69ffaa1edeba9d7408c67f2cface8c4113ffcbe9
Reviewed-by: Christian Tismer <tismer@stackless.com>