shiboken: Refactor generation of sequence access methods
Rename writeStdListWrapperMethods() to
writeDefaultSequenceMethods() since that is is more close to its
purpose. In the function, get the base container type.

Use std::advance instead of a loop to position the iterator.
This is specialized for random access iterators to perform an
addition and thus more efficient.

Use const_iterator in __getitem__ to prevent Qt containers from
detaching.

Task-number: PYSIDE-904
Change-Id: I4735f39193c4f4efa856440ecddbc48b3a5071ae
Reviewed-by: Christian Tismer <tismer@stackless.com>