Use pyside2-uic instead of uic for the loadUiType
Since we deploy the pyside2-uic wrapper inside the
bin/ directory of virtual environments, that takes care
of using the 'uic' binary we ship with the wheels,
which is located in site-packages/PySide2/.

The current implementation of loadUiType, runs 'uic -g python'
but for people without a Qt installation, it has no effect
since 'uic' is neither in PATH, nor in the system.

Using 'pyside2-uic' instead will solve this issue.

Task-number: PYSIDE-1223
Change-Id: I2c801a16c9ff4faa5cf8711bd718a634f017e592
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>