feature-select: Implement a selectable feature framework
This is the framework for selectable features.
There are no real features implemented.

Planned is a maximum of 8 features.
They are all implemented as a dummy for now.

The decision depends of the following setting at the beginning of
a module after PySide2 import:

    from __feature__ import <feature name>

For more info, see the Jira issue, section

    The Principle Of Selectable Features In PySide

Task-number: PYSIDE-1019
Change-Id: If355e9294b5c16090b39d30422a90ea9c8523390
Reviewed-by: Christian Tismer <tismer@stackless.com>