shiboken: optimize method override strings
The override strings in bindingmanager were based upon
pure C strings for a long time. There was some mild overhead
of creating PyObject all the time.

With the upcoming name mangling of Selectable Features,
this becomes even a bit more inefficient when strings also get
mangled. With Python strings this is just a dict lookup.

This patch uses the same machinery as `Shiboken::PyName`.

Task-number: PYSIDE-1019
Change-Id: I4eed7222371dadeed0e9ba98cc7970e726ffc044
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>