feature-select: allow snake_case instead of camelCase for methods
This is the implementation of the first of a series of dynamically
selectable features.

The decision depends of the following setting at the beginning of
a module after PySide2 import:

    from __feature__ import snake_case

For more info, see the Jira issue, section

    The Principle Of Selectable Features In PySide

The crucial problems that are now solved were:

- it is not sufficient to patch a type dict, instead the whole
  `tp_mro` must be walked to rename everything.

- tp_getattro must be changed for every existing type. This
  is done either in shiboken by a changed PyObject_GenericGetAttr
  or PyObject_SenericGetAttr, or in the generated tp_(get|set)attro
  functions.

An example is included in sources/pyside2/doc/tutorial/expenses.

Task-number: PYSIDE-1019
Change-Id: I5f103190be2c884b0b4ad806187f3fef8e6598c9
Reviewed-by: Christian Tismer <tismer@stackless.com>