PyName: finalize static strings, finally
Static strings are used in PyName and PyMagicName very much.
With the introduction of name mangling and the __feature__
selector, many thousands of static strings are created.

The destruction of static strings was activated and the
implementation was corrected. The principle was reverted to
a simple vector, because interned strings behave like normal
strings. The sharing of references happens transparently.

Change-Id: I915ab6585fe4f81e0d05b9da82842bb33e0bc938
Reviewed-by: Christian Tismer <tismer@stackless.com>