feature-select: fix crash in scriptable application
The feature selection framework had a very early static
initialization from the early stages of the development.

Normally, Python is initialized before PySide gets loaded.
In case of scriptable application, this might not be so.

Actually this static initialization was no longer necessary
and was removed, as the bug was hopefully as well.

Change-Id: I2c703c9cac14b6093d0c43c4bae94ff2b29c0640
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>