QQmlComponent: allow thread in constructors
The constructor of QQmlComponent uses qml threads,
and their usage was producing a deadlock when using along
Python classes for a urlInterceptor, mainly related to the
'allow-thread' story that we need to take care of.

Since the QQmlComponent required threads, we enable
all the constructors to allow them to avoid a deadlock.

Fixes: PYSIDE-1349
Change-Id: I86cb97b011f483cc4142bed6fae1ea525fa0db47
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>