feature-select: delay the feature switching
Feature switching was written rather naively and happened after
almost every module change which can occour very often.

This patch is much more careful with switching and uses an ignore
code to prevent switching when PySide was not imported at all.

A potential regression when a switch is set before some PySide
module is being imported and PepType_SOTP is zero was also solved.

Task-number: PYSIDE-1019
Change-Id: I24dec7b3e4d0b577f77262392ded0b8a2006b3cc
Reviewed-by: Christian Tismer <tismer@stackless.com>