signature: pass `self` directly from parser
The signature module took the info from the PyCFunction flags
for a long time to check if something is a function, method or
staticmethod.

It turned out that there are functions with multiple signatures
where the method/staticmethod info varies in the signatures.
This invalidated the PyCFunction flag usage.

Instead, we now compute that info directly from abstractmetalang.cpp
which has access to the correct info.

Fixes: PYSIDE-1328
Change-Id: I6ba7237efcc486de014184b1787d05d87bee5a5e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>