__feature__: ignore if `__name__` does not exist in a module
By the slow feature patch, a line got moved into a scope where it
is possible that the `__name__` attribute does not exist.

We supply a default value which solved the problem.

The problem with this bug was how to reproduce the bug.
The doc generator has probably never been run on macOS because
without the trick
    cp -r doc/CMakeFiles/qdoc.dir doc/CMakeFiles/qdoc/
the script does not work.

Change-Id: Ic1a65cfaaba5a5c1b0a8e9f390d70e70dfe38ebd
Fixes: PYSIDE-1368
Task-number: PYSIDE-1019
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>