Add compatibility with Nuitka
This patch is based upon the old PYSIDE-198 proposal.
It worked with a few changes on Python 3 with limited API disabled.

When enabling the Limited API, there were a lot of crashes. This was
due to the way we need to get around access to certain implementations.

This showed that the original patch was wrong in the expression of
bindingmanager.cpp

    bool isCompiled = !isMethod && Py_TYPE(method)->tp_call != nullptr;

After fixing this expression with

    bool isCompiled = !isMethod && Py_TYPE(method) != &PyCFunction_Type
                                && Py_TYPE(method)->tp_call != nullptr;

everything worked fine with the Limited API, too.

Fixes: PYSIDE-198
Task-number: PYSIDE-829
Change-Id: I4f887c639628041682052e90ba4c72aa98284e9e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>