testing: Ensure that build dir precedes site-packages
Before the big change
    "Fix running the PySide2 tests for Python 3.8/Windows"
the cmake paths were ordered in a way that a compilation
would not interfere with an installation.

With the new test layout, it suddenly became possible to
do a new compilation, but shiboken would not be loaded from
the compilation but use some version which is still in side-packages.

This patch fixes that so that the newly inserted directories
are guaranteed to come before site-packages.

Change-Id: Ib999664a798661533c4fe7bfab19458ce2a1d710
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>