shiboken2: Refactor property parsing
Store AbstractMetaType instead of TypeEntry in
QPropertySpec for more complete type information and
better parsing.

Introduce a struct TypeSystemProperty to the
typesystem which can be later populated from a newly
introduced typesystem XML element. Rewrite the parser
to first populate a TypeSystemProperty and convert it
to QPropertySpec with type lookup in a second step.

Change the parser algorithm to first look for any of
the READ/WRITE tokens and split the tokens leading up
to them into type name and property name in order to
parse complex types like

Q_PROPERTY(QList<QSpriteSheetItem *> sprites READ sprites..

correctly.

Task-number: PYSIDE-1019
Change-Id: I942bc659cc236d31041cdc22e26a82d270599033
Reviewed-by: Christian Tismer <tismer@stackless.com>