Signature: Revert SbkSpecial_Type_Ready to PyType_Ready
The early signature module tried to minimize the visible
changes to the code base. It replaced the `PyType_Ready` call
by a special version which did other things as well.

We replace that special call by a more intuitive function
`InitSignatureStrings` that does exactly that and nothing more.

The functionality of the module is unchanged.

Change-Id: Ic2f9cd29b0352f0a24daa55b01420c77d103c0b2
Task-number: PYSIDE-510
Reviewed-by: Christian Tismer <tismer@stackless.com>