Signature: Break the source into multiple files
The signature module has grown groups of rather unrelated
topics. In order to reduce the complexity, we break up
the source into a number of files.

Reason for this change was building support for selectable
features, which should not get lost in all the unrelated
helper functions.

Task-number: PYSIDE-510
Task-number: PYSIDE-1019
Change-Id: I8e22a91db1882f8c5428b8def13bf9f1cea431fb
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>