Feature-select: Implement signature-awareness of snake_case
After implementing selectable features, support from the signature
module was quite much missing. It was not clear for some time what
to do the best.

It turned out to have the smallest impact and runtime penalty
to use the Python parser output dictionaries and create copies
with snake case naming. That has almost no overhead.

Also, it was necessary to augment the internal map_dict with
snake_case versions. It may be possible to simplify that map_dict
further in another check-in.

Remaining is the problem of static properties. This will be tried
using the PySide Property objects which can be improved.

Change-Id: Ied83ccb197a3c15932c4202b5f1ade772416e17b
Task-number: PYSIDE-1019
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>