Fix leak in static string destruction
When checking a debug build, refcounting problems showed up.
It is unclear where references are lost, at the moment at the
strings "__div__" and "__idiv__".

It is possible that these strings were already in a destruction
list of another module. When turning the list into a set, the
problem vanishes, because we no longer know how much references
to expect. Because they are not negative, this seems to be ok.

Change-Id: I0bc59b83f43becf5ee5f5f19213fc46513041909
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>