Fix leak in static string destruction, addendum
When checking a debug build, refcounting problems showed up.
The reason was that somewhere the static strings were
treated with normal refcounting. This is easily possible if
some string is used in the interpreter: The string becomes
interned, maybe before it is used by us, and will be
deallocated without our knowledge.

We are turning the list of referenced strings into a set.
Also, the strings are made immortal by adding one extra count.
This way, our set does not need to cope with deletions and
will work either way.
The final cleanup removes that, anyway.

Change-Id: I192ea47f739cc785ea3e354cfb82c5fa5a6eac78
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>