feature_select: finish properties and postpone static properties
Static properties are not easy to support.
They need an extra property subtype and support in the meta-class.

A problem is that the implementation needs to go deeply into the
innards of things and need to use `_PyType_Lookup`.
That is hard to circumvent when the limited API is used.

Therefore, the current implementation will be it for 5.15 .
At most we might add a few missing properties through XML.

Task-number: PYSIDE-1019
Change-Id: I56a9eabe5f774c1ff04c149227e06318c8bf4f29
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>