__feature__: provide useful error message when feature is active
Features seem to work quite good so far, but they clearly need
much more testing.

One of the problems is that error messages are produced,
but the function name is not found and we get an ugly default
error message.

This is an efficient implementation that does all name mangling
in the C code. A helper dict avoids linear search for properties.

Task-number: PYSIDE-1019
Change-Id: Ic87c4a6e7dc2b2a251e809d6df0eb7fb9ca8021c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>