shiboken: Ensure that `seterror_argument` always gets a function arg
The call to the internal function `seterror_argument` is generated
by cppgenerator for all generated classes. The constructor is
special cased, so that for `SomeClass.__init__` only `SomeClass` is
generated.

For error handling, we should always see the failed function name,
regardless of any convention. This has become a real problem with the
new error handler written in C that can not easily deduce if a
function or class is passed.

Task-number: PYSIDE-1019
Change-Id: I073f8e866c6bbce5dd842ef4be5d6c6c5421db0f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>