shiboken6: Cache the modifications of AbstractMetaFunction
A lot of query functions depend on
AbstractMetaFunction::modifications(AbstractMetaClass*) which
goes on a search through the classes on each invocation,
depending on the class passed in. Cache the result in a flat list
of classes (which will typically contain
declaring/implementing/owner class or 0 for free functions).

Return modifications by const-ref and remove variables.

Change-Id: I8512b25d8c39f9dea5fc9141575902f89b4ad597
Reviewed-by: Christian Tismer <tismer@stackless.com>