shiboken6: Introduce a new Text stream
Add a new text stream wrapping QTextStream with built-in indent.
This removes the need to have INDENT as a separate member variable
or function parameter and to output it at the beginning of each
line.

It also makes it easier to factor out the write() functions since they
no longer need the indent.

Make the write* functions const where possible.

The doc generator is left as-is for the moment.
Rename the old Indentation to Indentation4 to avoid
conflicts.

Change-Id: I6866e7360df119859adeee567d2b18a37cb63a79
Reviewed-by: Christian Tismer <tismer@stackless.com>