cppgenerator: rework keyword handling regarding unknown.. Fixed!
The new keyword algorithm did fail again on Windows.
But this time it was a refcounting error that was only visible
in a debug build.

On Windows, it appeared as a negative refcount assertion.
On macOS in debug mode, it was a segfault instead.

The problem was `PyErr_Fetch()`, a function that uses borrowed
references, only. Forgot to respect that, sorry :-(

Task-number: PYSIDE-1305
Change-Id: I164668db8143729fa2fab0be1f61757690e13138
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 76a72315de6cd9ea2f76109a3896d95db4227f64)