shiboken6: Rewite the graph sorter
Change the Graph used for dependency sorting from a graph
using integer node numbers to a template taking the Node value,
relying on operator==() (and its qDebug() operators).

The mapping of node to indexes can then be removed from the client code,
leading to a significant simplification.

As a side effect, this fixes undefined behavior of the overload
sorter in conjunction with reverse binary operators. It was not handling
overloads of the same decisor argument type correctly, leading to
graphs with duplicated node types for them.

Rewrite the toposort test to be data driven.

Change-Id: Idbe896dc81d7272099db6dab3d2673643fc17401
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 7626f04ac855a56d4b364eac7b8350bdd561fef0)