shiboken6: Refactor base class retrieval
AbstractMetaClass has a list of base class names and a list
of AbstractMetaClass* base instances.

AbstractMetaBuilder populates the instance list from the
names in setupInheritance() and sets a flag.

In a few places in AbstractMetaBuilder, the base class
instances are needed before this has been completed.
For this purpose, the helper
AbstractMetaBuilder::getBaseClasses() is needed.

Replace the set of classes m_setupInheritanceDone
by a flag on AbstractMetaClass. This allows for adding
an assert to AbstractMetaClass::baseClasses() ensuring
setupInheritance() has been called and taking a shortcut
in AbstractMetaBuilder::getBaseClasses().

In addition, classesTopologicalSorted() can use
AbstractMetaClass::baseClasses() instead of searching with
AbstractMetaBuilder::getBaseClasses() and can thus be
made static.

Change-Id: Iaf8209b6f6534ad91a96970a56c1e86cce054922
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit e99ac18b134469854dff05fd3bb61158f9d9ccf9)
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>