qApp: fix flag handling in Python 3.8+ and a Python 3.9 issue
This is an old problem that was solved for Python 2.7.
From Python 3.8 on, the behavior is the same with Python 3.

The fix finally was to extend a Python 2.7 patch to Python 3
as well. See the Jira issue for details.

Other little changes:

This patch includes also a small patch that was mentioned as
necessary for Python 3.9: Python issue 40217. I have seen no
effect of this change yet but applied the patch, anyway.

When searching for a solution of this problem, a lot of time
was spent debugging qapp_macro.cpp, although it was error-free.
As a side effect for better understanding, the variables were
renamed to more common style.

These more independent changes could have got their own check-in,
but the complication of a pick-to and the small code size wasn't
worth the effort.

Change-Id: I7638f1a711315b4678af6b7389265b905c6404a1
Fixes: PYSIDE-1447
Reviewed-by: Dmitry Shachnev <mitya57@gmail.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit c1f776570adbe4f9aa21aa818a82f5ebd1258a76)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>