Fix typo while setting LLVM_INSTALL_DIR
The original commit was done for 6.1 and cherry-picked to 5.15. The
problem is that 6.1 is missing some of the older CI configs there are
in 5.15.
This change will enable libclang v10.0 only for such CI configs, which
have the correct libclang installed. Other configs will keep using
the one there is installed on VM.

Change-Id: Ifd2a461b632945586fd4e3fa16a5c1462a5fb493
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>