feature: delete setter after creating property only if standard
Some features like QWidget.size are defined by Qt with a
non-standard setter name. For size, the standard setter name
would be setSize, and in the property creation process
this setter would be deleted.

We changed rules in this way:
If a setter name is non-standard (like resize), the setter will
not be removed but still can be used.

Actually it would make more sense if "size" was a read-only
property.

Task-number: PYSIDE-1019
Change-Id: I9ded7e9c1dbd2932aa4c5616385b90ed673bfaee
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit e675a51ce48875800c1f9137cedd49144c62af58)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>